+++ src/historical.cpp
@@ -96,6 +96,21 @@ T FromCheckedAtString(const std::string& endpoint, const nlohmann::json& json,
 }
 
 template <typename T>
+T FromCheckedAtStringOrNull(const std::string& endpoint,
+                            const nlohmann::json& json, const std::string& key,
+                            T null_value) {
+  const auto& val_json = ::CheckedAt(endpoint, json, key);
+  if (val_json.is_null()) {
+    return null_value;
+  }
+  if (val_json.is_string()) {
+    return databento::FromString<T>(val_json);
+  }
+  throw JsonResponseError::TypeMismatch(endpoint, key + " null or string",
+                                        val_json);
+}
+
+template <typename T>
 T ParseAt(const std::string& endpoint, const nlohmann::json& json,
           const std::string& key);
 
@@ -200,13 +215,14 @@ databento::BatchJob Parse(const std::string& endpoint,
   res.end = ParseAt<std::string>(endpoint, json, "end");
   res.limit = ParseAt<std::size_t>(endpoint, json, "limit");
   res.encoding = FromCheckedAtString<Encoding>(endpoint, json, "encoding");
-  res.compression =
-      FromCheckedAtString<Compression>(endpoint, json, "compression");
-  res.split_duration =
-      FromCheckedAtString<SplitDuration>(endpoint, json, "split_duration");
+  res.compression = FromCheckedAtStringOrNull<Compression>(
+      endpoint, json, "compression", Compression::None);
+  res.split_duration = FromCheckedAtStringOrNull<SplitDuration>(
+      endpoint, json, "split_duration", SplitDuration::None);
   res.split_size = ParseAt<std::size_t>(endpoint, json, "split_size");
   res.split_symbols = ParseAt<bool>(endpoint, json, "split_symbols");
-  res.packaging = FromCheckedAtString<Packaging>(endpoint, json, "packaging");
+  res.packaging = FromCheckedAtStringOrNull<Packaging>(
+      endpoint, json, "packaging", Packaging::None);
   res.delivery = FromCheckedAtString<Delivery>(endpoint, json, "delivery");
   res.record_count = ParseAt<std::size_t>(endpoint, json, "record_count");
   res.billed_size = ParseAt<std::size_t>(endpoint, json, "billed_size");
@@ -258,30 +274,32 @@ databento::BatchJob Historical::BatchSubmitJob(
     const std::string& dataset, UnixNanos start, UnixNanos end,
     const std::vector<std::string>& symbols, Schema schema) {
   return this->BatchSubmitJob(dataset, start, end, symbols, schema,
-                              SplitDuration::Day, {}, Packaging::None,
-                              Delivery::Download, kDefaultSTypeIn,
-                              kDefaultSTypeOut, {});
+                              Compression::Zstd, SplitDuration::Day, {},
+                              Packaging::None, Delivery::Download,
+                              kDefaultSTypeIn, kDefaultSTypeOut, {});
 }
 databento::BatchJob Historical::BatchSubmitJob(
     const std::string& dataset, const std::string& start,
     const std::string& end, const std::vector<std::string>& symbols,
     Schema schema) {
   return this->BatchSubmitJob(dataset, start, end, symbols, schema,
-                              SplitDuration::Day, {}, Packaging::None,
-                              Delivery::Download, kDefaultSTypeIn,
-                              kDefaultSTypeOut, {});
+                              Compression::Zstd, SplitDuration::Day, {},
+                              Packaging::None, Delivery::Download,
+                              kDefaultSTypeIn, kDefaultSTypeOut, {});
 }
 databento::BatchJob Historical::BatchSubmitJob(
     const std::string& dataset, UnixNanos start, UnixNanos end,
     const std::vector<std::string>& symbols, Schema schema,
-    SplitDuration split_duration, std::size_t split_size, Packaging packaging,
-    Delivery delivery, SType stype_in, SType stype_out, std::size_t limit) {
+    Compression compression, SplitDuration split_duration,
+    std::size_t split_size, Packaging packaging, Delivery delivery,
+    SType stype_in, SType stype_out, std::size_t limit) {
   httplib::Params params{
       {"dataset", dataset},
       {"start", ToString(start)},
       {"end", ToString(end)},
       {"symbols", JoinSymbolStrings(kBatchSubmitJobEndpoint, symbols)},
       {"schema", ToString(schema)},
+      {"compression", ToString(compression)},
       {"encoding", "dbn"},
       {"split_duration", ToString(split_duration)},
       {"packaging", ToString(packaging)},
@@ -295,15 +313,16 @@ databento::BatchJob Historical::BatchSubmitJob(
 databento::BatchJob Historical::BatchSubmitJob(
     const std::string& dataset, const std::string& start,
     const std::string& end, const std::vector<std::string>& symbols,
-    Schema schema, SplitDuration split_duration, std::size_t split_size,
-    Packaging packaging, Delivery delivery, SType stype_in, SType stype_out,
-    std::size_t limit) {
+    Schema schema, Compression compression, SplitDuration split_duration,
+    std::size_t split_size, Packaging packaging, Delivery delivery,
+    SType stype_in, SType stype_out, std::size_t limit) {
   httplib::Params params{
       {"dataset", dataset},
       {"start", start},
       {"end", end},
       {"symbols", JoinSymbolStrings(kBatchSubmitJobEndpoint, symbols)},
       {"schema", ToString(schema)},
+      {"compression", ToString(compression)},
       {"encoding", "dbn"},
       {"split_duration", ToString(split_duration)},
       {"packaging", ToString(packaging)},